Moved template types to Utility
