+++ Include/Misc/Utility.h
@@ -15,12 +15,14 @@ namespace Rift
 		std::swap(a, b);
 	}
 
+
 	// Forward arg as movable
 	template <typename T>
 	constexpr std::remove_reference_t<T>&& Move(T&& arg) noexcept
 	{
 		return static_cast<std::remove_reference_t<T>&&>(arg);
 	}
+
 
 	// Forward an lvalue as either an lvalue or an rvalue
 	template <typename T>
@@ -29,6 +31,7 @@ namespace Rift
 		return static_cast<T&&>(arg);
 	}
 
+
 	// Forward an rvalue as an rvalue
 	template <class T>
 	constexpr T&& Forward(std::remove_reference_t<T>&& arg) noexcept
@@ -36,4 +39,40 @@ namespace Rift
 		static_assert(!std::is_lvalue_reference_v<T>, "Bad Forward call");
 		return static_cast<T&&>(arg);
 	}
+
+
+	template <typename Predicate>
+	class ReversePredicate
+	{
+		const Predicate& predicate;
+
+	public:
+		ReversePredicate(const Predicate& predicate) : predicate(predicate) {}
+
+		template <typename T>
+		bool operator()(T&& A, T&& B) const
+		{
+			return predicate(Forward<T>(B), Forward<T>(A));
+		}
+	};
+
+
+	struct Less
+	{
+		template <typename T>
+		bool operator()(T&& A, T&& B) const
+		{
+			return Forward<T>(A) < Forward<T>(B);
+		}
+	};
+
+
+	struct More
+	{
+		template <typename T>
+		bool operator()(T&& A, T&& B) const
+		{
+			return Forward<T>(A) > Forward<T>(B);
+		}
+	};
 }    // namespace Rift
